refactor(media): move the capability descriptors into core
Nothing about describing a capability is DOM-specific, so the descriptors sit
beside the model that defines them. Three of them only lived under `dom`
because they had been narrowed to `TimeRanges` and `TextTrackList`; the
framework-neutral contracts they already implement say the same thing, and
every consumer was reading them through those contracts anyway.

Core's tsconfig has no DOM lib, so the location now enforces what the
descriptors were meant to be: inert data any adapter can compose or inspect,
with no reference to a host or a browser. It also settles a packaging gap,
since the descriptors ship with the package's main entry rather than through a
subpath the build never declared.
