+++ packages/media/src/core/capabilities/autoplay.ts
@@ -1,5 +1,5 @@
-import { defineMediaCapability } from '../../core/capability';
-import type { MediaAutoplayCapability } from '../../core/types';
+import { defineMediaCapability } from '../capability';
+import type { MediaAutoplayCapability } from '../types';
 
 /** Starting playback without a gesture. */
 export const autoplayCapability = defineMediaCapability<MediaAutoplayCapability>()({
