+++ packages/media/src/core/capabilities/controls.ts
@@ -1,5 +1,5 @@
-import { defineMediaCapability } from '../../core/capability';
-import type { MediaControlsCapability } from '../../core/types';
+import { defineMediaCapability } from '../capability';
+import type { MediaControlsCapability } from '../types';
 
 /** The media's own native controls. */
 export const controlsCapability = defineMediaCapability<MediaControlsCapability>()({
