+++ packages/media/src/core/capabilities/error.ts
@@ -1,5 +1,5 @@
-import { defineMediaCapability } from '../../core/capability';
-import type { MediaErrorCapability } from '../../core/types';
+import { defineMediaCapability } from '../capability';
+import type { MediaErrorCapability } from '../types';
 
 /** Reporting a fatal playback failure. */
 export const errorCapability = defineMediaCapability<MediaErrorCapability>()({
