+++ packages/media/src/core/capabilities/live.ts
@@ -1,5 +1,5 @@
-import { defineMediaCapability } from '../../core/capability';
-import type { MediaLiveCapability } from '../../core/types';
+import { defineMediaCapability } from '../capability';
+import type { MediaLiveCapability } from '../types';
 
 /** Describing the live window a stream exposes. */
 export const liveCapability = defineMediaCapability<MediaLiveCapability>()({
