+++ packages/media/src/core/capabilities/pause.ts
@@ -1,5 +1,5 @@
-import { defineMediaCapability } from '../../core/capability';
-import type { MediaPauseCapability } from '../../core/types';
+import { defineMediaCapability } from '../capability';
+import type { MediaPauseCapability } from '../types';
 
 /** Suspending playback. A media that plays through once, unstoppably, composes `playback` without this. */
 export const pauseCapability = defineMediaCapability<MediaPauseCapability>()({
