+++ packages/media/src/core/capabilities/playback-rate.ts
@@ -1,5 +1,5 @@
-import { defineMediaCapability } from '../../core/capability';
-import type { MediaPlaybackRateCapability } from '../../core/types';
+import { defineMediaCapability } from '../capability';
+import type { MediaPlaybackRateCapability } from '../types';
 
 /** Playing faster or slower than real time. */
 export const playbackRateCapability = defineMediaCapability<MediaPlaybackRateCapability>()({
