+++ packages/media/src/core/capabilities/playback.ts
@@ -1,5 +1,5 @@
-import { defineMediaCapability } from '../../core/capability';
-import type { MediaPlaybackCapability } from '../../core/types';
+import { defineMediaCapability } from '../capability';
+import type { MediaPlaybackCapability } from '../types';
 
 /** Starting playback. The one capability every media has; without it there is nothing to host. */
 export const playbackCapability = defineMediaCapability<MediaPlaybackCapability>()({
