+++ packages/media/src/core/capabilities/poster.ts
@@ -1,5 +1,5 @@
-import { defineMediaCapability } from '../../core/capability';
-import type { MediaPosterCapability } from '../../core/types';
+import { defineMediaCapability } from '../capability';
+import type { MediaPosterCapability } from '../types';
 
 /** A still image standing in for the content before playback. */
 export const posterCapability = defineMediaCapability<MediaPosterCapability>()({
