+++ packages/media/src/core/capabilities/volume.ts
@@ -1,5 +1,5 @@
-import { defineMediaCapability } from '../../core/capability';
-import type { MediaVolumeCapability } from '../../core/types';
+import { defineMediaCapability } from '../capability';
+import type { MediaVolumeCapability } from '../types';
 
 /**
  * Volume forwarding.
