--- packages/media/src/dom/capabilities/played.ts
@@ -1,17 +1,0 @@
-import { defineMediaCapability } from '../../core/capability';
-import { EMPTY_TIME_RANGES } from '../../core/constants';
-import type { MediaPlayedCapability } from '../../core/types';
-
-/** DOM-shaped {@link MediaPlayedCapability}: browser hosts hand back real `TimeRanges`. */
-export interface HTMLMediaPlayedCapability extends MediaPlayedCapability {
-  readonly played: TimeRanges;
-}
-
-/** Reporting which parts of the timeline the viewer has actually watched. */
-export const playedCapability = defineMediaCapability<HTMLMediaPlayedCapability>()({
-  name: 'played',
-  events: [],
-  props: {
-    played: { fallback: EMPTY_TIME_RANGES as TimeRanges, readonly: true },
-  },
-});
